src/backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas
scikit-learn
joblib
azure-storage-blob
cryptography==41.0.7
requests-toolbelt
mysql-connector-python
numpy
